+++ integrations/intercom/serializers/intercom-attributes.js
@@ -1,10 +1,8 @@
 'use strict';
 var JSONAPISerializer = require('jsonapi-serializer').Serializer;
-var StringsUtil = require('../../../utils/strings');
 
 function IntercomAttributesSerializer(attributes, collectionName, meta) {
-  var type = StringsUtil.camelCaseToDashed(collectionName) +
-                                           '-intercom-attributes';
+  var type = collectionName + '_intercom_attributes';
 
   return new JSONAPISerializer(type, attributes, {
     attributes: ['session_count', 'last_seen_ip', 'created_at', 'updated_at',