+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [Unreleased]
+- Collections Names - Collection names are now based on the model name whatever the ORM is.
 
 ## RELEASE 1.5.3 - 2017-11-27
 ### Added